+++ src/criteria/default/referrer.js
@@ -1,4 +1,4 @@
-import { setMatchingAttribute, setMatchingFunction } from '../utils';
+import { setMatchingAttribute } from '../utils';
 
 const matchingAttribute = ( { store } ) => {
 	const value = document.referrer
@@ -13,18 +13,3 @@ const matchingAttribute = ( { store } ) => {
 
 setMatchingAttribute( 'sources_to_match', matchingAttribute );
 setMatchingAttribute( 'sources_to_exclude', matchingAttribute );
-
-setMatchingFunction( 'sources_to_exclude', ( config, { store } ) => {
-	let list = config.value;
-	if ( typeof list === 'string' ) {
-		list = config.value.split( ',' ).map( item => item.trim() );
-	}
-	if ( ! Array.isArray( list ) || ! list.length ) {
-		return true;
-	}
-	const value = store.get( 'referrer' );
-	if ( ! value || ! list.includes( value ) ) {
-		return true;
-	}
-	return false;
-} );
